[Fix #4081] Modularize gRPC (#4160)
* Modularize gRPC


* Add default ctor for WorkflowRPCHandlerGenerator


* Apply impsort plugin


* Remove reference of grpc stuff from ServerlessWorkflowAssetsProcessor


* Move grpc and openapi stuff to the respective modules


* Apply impsort plugin


* Remove kogito.rpc configurations from kogito-serverless-workflow-extension


---------
