+++ quarkus/addons/grpc/deployment/src/main/resources/META-INF/services/org.kie.kogito.quarkus.serverless.workflow.deployment.livereload.LiveReloadableCodeGenProvider
@@ -17,4 +17,4 @@
 # under the License.
 #
 
-org.kie.kogito.quarkus.serverless.workflow.deployment.livereload.LiveReloadableAsyncApiGeneratorStreamCodeGen
+org.kie.kogito.quarkus.serverless.workflow.rpc.LiveReloadableWorkflowRPCCodeGenProvider
